final -- initial version of the python package docker-compose-help
