+++ .gitignore
@@ -5,3 +5,4 @@ dist
 *.pyc
 *$py.class
 *.egg-info
+*setup.py
